Data type tuning in PrimeSievePY
A couple of minor tweaks. Not all are significant, but combine this with PyPy, and the program is 50x times faster. PyPy alone was 16x faster. Haven't made it compatible with MicroPython viper, which tends to be comparable in speed. 

// instead of /: floor (integer) division. No need to call int() on the results. 

bytearray instead of list: far more compact. 

slots: faster attribute access. 

Of course, this isn't normally a Python task, and this is not a particularly efficient or idiomatic implementation in Python; removing the sanity check in ClearBit alone can speed it up 19%. 
For instance, one might use the bitarray module (which is not in the standard library).